Consolidates region and zonal snapshots to single resource.
